LayoutTests/fast/dom/HTMLOutputElement/htmloutputelement-children-removed.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test that a live list is updated after the child of an HTMLOutputElement was removed.');
</script>
<output id="parentOutput"><span class="inner" id="first"></span><span class="inner" id="second"></span></output>
<script>
    var list = document.getElementsByClassName("inner");
    debug("The number of spans before removal is " + list.length);
    var beforeLength = list.length;
    document.getElementById("parentOutput").removeChild(document.getElementById("first"));
    debug("The number of spans after removal is " + list.length);
    shouldBeTrue('list.length == beforeLength - 1');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
